add save residue index feature
Residue indices from structure files can now be optionally saved in the db output. The label seq id is taken, not auth seq id. The option is used in createdb; in createdimerdb, createsubdb, and createinterfacedb, the index db is modified if it exists. Use option --save-res-index, output is stored in <db>_id